First Commit, Stable and Dev version
This is basically a commit of all the files for this particular project
on my computer for the purpose of sharing with others.  The stable
version should always work...the dev one will frequently be broken or
have bugs as I try to add new features.  Once a feature is added and
works properly, I will add a new copy of the stable version to include
the feature.